refactor: remove batch-era mDNS machinery replaced by StreamLAN
Deletes the pre-StreamLAN batch/continuous discovery paths across all
three platforms now that CollectLAN/StreamLAN (stream.go) fully own LAN
discovery:

- discovery.go: DiscoverLANContinuous (zero callers since Task 10).
- discovery_darwin.go: discoverLAN, discoverLANContinuous, dnssdResolve,
  deviceFromBrowse, setAssetID. Keeps the live init() annotator,
  browseResult, darwinCachedInterfaceLinkSpeed, isValidHostnameLabel.
- discovery_linux.go: deleted entirely (discoverLAN, discoverLANAvahi,
  parseAvahiResolveLine, avahiUnescape, parseAvahiTXT, hasAvahiBrowse,
  discoverLANMDNS, lanDeviceFromMDNSEntry, queryInterface,
  discoverLANContinuous, setAssetID, setMeshFields) — the avahi-browse
  child-process path dies with it; the live path is avahi_dbus_linux.go
  (D-Bus, no exec) with backend_hashicorp.go as fallback.
- discovery_windows.go: deleted entirely (discoverLAN and its full
  support chain — every symbol in the file was batch-path-only).
- dnssd_darwin.go: dnssdBrowse (only caller was discoverLAN; the cgo
  binding it wrapped, dnssdBrowseStream/dnssdResolveInstance, is
  unaffected and already covered by mdns_darwin_test.go).
- usb_connection.go: appendPreferredLANDevice and its dead scoring chain
  (preferDiscoveredLANDevice, lanDeviceDiscoveryScore,
  isRoutableLANAddress, isIPv4LANAddress) — its only callers were the
  deleted discoverLAN batch functions.

Test files for deleted production code are deleted with it. Tests that
covered still-live functions incidentally defined in deleted files move
to their new home: TestParseMDNSInfoFields → backend_hashicorp_test.go,
the parseNetAdapterJSON tests → new ethernet_windows_test.go. Stale
comments pointing at now-deleted symbols/line numbers are updated.

Verified: go test ./go/internal/... -race (native + dockerized linux)
green, GOOS=windows/linux go vet clean, gofmt clean, and
`exec.Command`/`exec.LookPath` no longer appear anywhere in the mDNS
browse path (mdns*.go, dnssd*.go, discovery_linux.go) — the
Bluetooth/Ethernet/USB/serial shell-outs elsewhere in the package are
unrelated and untouched.
