+++ go/internal/shared/discovery/avahi_dbus_linux_test.go
@@ -135,7 +135,7 @@ func TestDecodeResolveReply(t *testing.T) {
 		if svc.InstanceName != "wendyos-foo" {
 			t.Errorf("InstanceName = %q, want %q", svc.InstanceName, "wendyos-foo")
 		}
-		// Trailing dot stripped, same as parseAvahiResolveLine (discovery_linux.go).
+		// Trailing dot stripped.
 		if svc.Hostname != "wendyos-foo.local" {
 			t.Errorf("Hostname = %q, want %q (trailing dot stripped)", svc.Hostname, "wendyos-foo.local")
 		}
